chore: Move github_user module to sub module
