+++ src/github_user/domain/entities/group.py
